기본이미지 삭제 불가 구현

+++ src/components/crud/FeedItem.jsx
@@ -45,7 +45,13 @@ function FeedItem() {
   ///사진 삭제 구현 수정중....
   const deleteImg = async () => {
     try {
+      const defaultImg = ref(storage, `/defaultImg/background.png`);
+      if (defaultImg) {
+        alert("기본 이미지입니다.");
+        return;
+      }
       const desertRef = ref(storage, img);
+
       await deleteObject(desertRef);
       alert("사진이 삭제됐습니다. 마저 수정을 완료해주세요");
     } catch (error) {
